+++ gpu/allocation_gpu.cuh
@@ -7,18 +7,18 @@
 
 #include "../utils/allocation.h"
 
-DTYPE* allocate1DArrayDevice(int n);
+DTYPE* allocate1DArrayDevice(size_t n);
 
 
-void copy1DFromHostToDevice(DTYPE* host, DTYPE* device, int n);
+void copy1DFromHostToDevice(DTYPE* host, DTYPE* device, size_t n);
 
-void copy2DFromHostToDevice(DTYPE** host, DTYPE* device, int n, int m);
+void copy2DFromHostToDevice(DTYPE** host, DTYPE* device, size_t n, size_t m);
 
-void copy1DFromDeviceToHost(DTYPE* device, DTYPE* host, int n);
+void copy1DFromDeviceToHost(DTYPE* device, DTYPE* host, size_t n);
 
-void copy2DFromDeviceToHost(DTYPE* device, DTYPE** host, int n, int m);
+void copy2DFromDeviceToHost(DTYPE* device, DTYPE** host, size_t n, size_t m);
 
-DTYPE* copy1DArrayDevice(int n, DTYPE* old);
+DTYPE* copy1DArrayDevice(size_t n, DTYPE* old);
 
 void free1DArrayDevice(DTYPE* device);
